feat : 수학공식을 몰랐나보다

+++ src/main/java/org/example/math/Q2166.java
@@ -7,7 +7,7 @@ public class Q2166 {
     static int N ;
     static int[][] index;
     public static void main(String[] args) {
-        int a=0,b=0;
+        long a=0,b=0;
         Scanner scan = new Scanner(System.in);
         N = scan.nextInt();
         index = new int[N][2];
@@ -19,9 +19,9 @@ public class Q2166 {
             index[i][1] = y;
         }
 
-        for (int i=0;i<N-1;i++){
-            a += index[i][0] * index[i+1][1];
-            b += index[i][1] * index[i+1][0];
+        for (int i = 0; i < N-1; i++) {
+            a += (long) index[i][0] * index[i + 1][1];
+            b += (long) index[i][1] * index[i + 1][0];
         }
 
         double area = Math.abs(a - b) / 2.0;
